+++ tcex/sessions/external_session.py
@@ -4,33 +4,50 @@ import urllib3
 from requests import Session, adapters
 from urllib3.util.retry import Retry
 
+from ..utils import Utils
+
 # disable ssl warning message
 urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
 
 
 class ExternalSession(Session):
-    """ThreatConnect REST API Requests Session for external requests"""
+    """ThreatConnect REST API Requests Session for external requests
 
-    def __init__(self, tcex):
+    Args:
+        logger (Logger): An instance of Logger
+    """
+
+    __attrs__ = [
+        'adapters',
+        'auth',
+        'cert',
+        'cookies',
+        'headers',
+        'hooks',
+        'max_redirects',
+        'proxies',
+        'params',
+        'stream',
+        'verify',
+        'trust_env',
+        # custom attrs
+        '_base_url',
+        '_mask_headers',
+        '_mask_patterns',
+        'log',
+        'utils',
+    ]
+
+    def __init__(self, logger):
         """Initialize the Class properties."""
         super().__init__()
-        self.tcex = tcex
+        self.log = logger
 
         # properties
         self._base_url = None
         self._mask_headers = True
         self._mask_patterns = None
-
-        # Update User-Agent
-        self.headers.update({'User-Agent': 'TcEx'})
-
-        # Set Proxy
-        if self.tcex.default_args.tc_proxy_external:
-            self.proxies = self.tcex.proxies
-            self.tcex.log.info(
-                'Using proxy server for external connectivity '
-                f'({self.tcex.default_args.tc_proxy_host}:{self.tcex.default_args.tc_proxy_port}).'
-            )
+        self.utils = Utils()
 
         # Add default Retry
         self.retry()
@@ -72,16 +89,16 @@ class ExternalSession(Session):
         response = super().request(method, url, **kwargs)
 
         # APP-79 - adding logging of request as curl commands
-        self.tcex.log.debug(
-            self.tcex.utils.requests_to_curl(
+        self.log.debug(
+            self.utils.requests_to_curl(
                 response.request,
                 mask_headers=self.mask_headers,
                 mask_patterns=self.mask_patterns,
                 verify=self.verify,
             )
         )
-        self.tcex.log.debug(f'request url: {response.request.url}')
-        self.tcex.log.debug(f'status_code: {response.status_code}')
+        self.log.debug(f'request url: {response.request.url}')
+        self.log.debug(f'status_code: {response.status_code}')
 
         return response